fix(map): split MapScreen / MapChip into lazy dispatchers
MapLibre's MapView.js calls `requireNativeComponent('MLRNMapView')` at
module-load time (line 439 of the installed v10.2). In Expo Go — where
the native module isn't registered — that call throws an Invariant
Violation the moment anything imports `@maplibre/maplibre-react-native`.

The previous `isMapNativeAvailable()` guards inside the components ran
after the import had already crashed, so they never fired. The only
reliable fix is to keep MapLibre out of the statically-imported module
graph when native isn't available — React.lazy does exactly that
because it defers the factory until render time.

Refactor:
  - src/constants/mapStyles.ts  (new) — STYLE_ANCIENT / STYLE_MODERN;
    a MapLibre-free module so dispatchers can share the URLs.
  - src/screens/MapScreen.tsx   (rewritten) — thin dispatcher. Exports
    buildPlaceToStoriesMap + the style URLs. If MapLibre native is
    missing, renders MapUnavailableCard; otherwise React.lazy-loads
    MapScreenNative inside a Suspense boundary.
  - src/screens/MapScreenNative.tsx  (renamed from the old MapScreen) —
    still wraps the full MapLibre implementation, minus the now-redundant
    in-component guard and the withErrorBoundary wrapper (both moved up
    to the dispatcher).
  - src/components/map/MapChip.tsx   (rewritten) — dispatcher that
    returns null when native is missing and lazy-loads MapChipNative
    otherwise.
  - src/components/map/MapChipNative.tsx  (renamed) — MapLibre impl.
  - ChapterVerseList keeps `import { MapChip } from './map/MapChip'` —
    dispatcher transparently wraps the native component.

Tests:
  - MapChip test now imports MapChipNative directly (jest can't resolve
    React.lazy's dynamic import without --experimental-vm-modules); new
    MapChipDispatcher.test covers the null-on-Expo-Go gate.
  - MapScreen test points at MapScreenNative; new MapScreenDispatcher
    test covers the MapUnavailableCard branch.

429 / 429 Jest suites pass; npx tsc --noEmit clean.
